+++ tests/testapp/urls_default.py
@@ -1,15 +1,11 @@
-try:
-    from django.urls import re_path as url
-except ImportError:
-    from django.conf.urls import url
-from django.conf.urls import include
+from django.urls import include, path
 from django.contrib.auth.views import LoginView, LogoutView
 
 
 urlpatterns = [
-    url(r'^', include('django_comments.urls')),
+    path('', include('django_comments.urls')),
 
     # Provide the auth system login and logout views
-    url(r'^accounts/login/$', LoginView.as_view(template_name='login.html')),
-    url(r'^accounts/logout/$', LogoutView.as_view()),
+    path('accounts/login/', LoginView.as_view(template_name='login.html')),
+    path('accounts/logout/', LogoutView.as_view()),
 ]